// src/app/api-cinekeep/model/episodeRatingResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EpisodeRatingResponse { 
    id?: number;
    seriesTmdbId?: number;
    seriesTitle?: string;
    seriesPosterPath?: string;
    seasonNumber?: number;
    episodeNumber?: number;
    episodeName?: string;
    stillPath?: string;
    airDate?: string;
    value?: number;
    ratedAt?: string;
    updatedAt?: string;
}