// src/app/api-cinekeep/model/pageResponseEpisodeRatingResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EpisodeRatingResponse } from './episodeRatingResponse';


export interface PageResponseEpisodeRatingResponse { 
    content?: Array<EpisodeRatingResponse>;
    page?: number;
    size?: number;
    totalElements?: number;
    totalPages?: number;
}