// src/app/api-cinekeep/model/pageResponseRatingResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RatingResponse } from './ratingResponse';


export interface PageResponseRatingResponse { 
    content?: Array<RatingResponse>;
    page?: number;
    size?: number;
    totalElements?: number;
    totalPages?: number;
}